requirements.txt
PySide6~=6.6.1
psutil~=5.9.7
py7zr~=0.20.8
xxtea~=3.2.0
requests~=2.31.0
pycryptodome==3.20.0
pillow~=10.2.0
ffmpeg-python